% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{confirm_modal}
\alias{confirm_modal}
\title{confirm_modal}
\usage{
confirm_modal(
  session,
  id,
  title,
  message,
  confirm_id,
  confirm_label = "Confirm",
  danger = FALSE
)
}
\arguments{
\item{id}{ID}

\item{title}{Title of the modal}

\item{message}{Modal message}

\item{confirm_label}{Confirmation Label}
}
\value{
A modal dialogue box
}
\description{
Create a reusable confirmation modal dialog
}
