% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotwidget.R
\name{hotwidget}
\alias{hotwidget}
\title{hotwidget htmlwidget wrapper}
\usage{
hotwidget(
  data = NULL,
  message = NULL,
  width = NULL,
  height = NULL,
  options = list()
)
}
\arguments{
\item{data}{A data.frame to render in the widget (or `message` for backwards compatibility).}

\item{width, height}{Widget width/height.}

\item{options}{List of widget options (colClasses, readOnly, etc).}
}
\description{
hotwidget htmlwidget wrapper
}
