% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{mod_table_server}
\alias{mod_table_server}
\title{mod_table_server}
\usage{
mod_table_server(id, store)
}
\arguments{
\item{id}{module id}

\item{store}{R6 DataStore instance}
}
\description{
Module server for editable table
}
