% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{summary_block}
\alias{summary_block}
\title{summary_block}
\usage{
summary_block(ns, title, value_ui)
}
\arguments{
\item{ns}{Character, namespace}

\item{title}{Character, title of summary block}

\item{value_ui}{UI output code block}
}
\value{
A \code{shiny.tag} object containing the summary block UI.
}
\description{
Summary value UI block
}
