% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStore.R
\name{DataStore}
\alias{DataStore}
\title{initialize}
\description{
An R6 class intended to connect to a bundled DuckDB DB (mtcars.duckdb) and
expose data, original, update_cell, revert, and summary methods.
}
\details{
DataStore R6 class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataStore-new}{\code{DataStore$new()}}
\item \href{#method-DataStore-save_to_db}{\code{DataStore$save_to_db()}}
\item \href{#method-DataStore-update_cell}{\code{DataStore$update_cell()}}
\item \href{#method-DataStore-save_snapshot}{\code{DataStore$save_snapshot()}}
\item \href{#method-DataStore-summary}{\code{DataStore$summary()}}
\item \href{#method-DataStore-disconnect}{\code{DataStore$disconnect()}}
\item \href{#method-DataStore-clone}{\code{DataStore$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStore-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataStore-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStore$new(db_path = NULL, table = "mtcars", read_only = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db_path}}{C character string of data base path}

\item{\code{table}}{Character string of table name}

\item{\code{read_only}}{Logical}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStore-save_to_db"></a>}}
\if{latex}{\out{\hypertarget{method-DataStore-save_to_db}{}}}
\subsection{Method \code{save_to_db()}}{
Method to overwrite the data with most recent changes back to database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStore$save_to_db()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStore-update_cell"></a>}}
\if{latex}{\out{\hypertarget{method-DataStore-update_cell}{}}}
\subsection{Method \code{update_cell()}}{
Method to update a single cell
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStore$update_cell(row, col, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStore-save_snapshot"></a>}}
\if{latex}{\out{\hypertarget{method-DataStore-save_snapshot}{}}}
\subsection{Method \code{save_snapshot()}}{
Method to revert back to original data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStore$save_snapshot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStore-summary"></a>}}
\if{latex}{\out{\hypertarget{method-DataStore-summary}{}}}
\subsection{Method \code{summary()}}{
Method to generate simple summary#'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStore$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStore-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-DataStore-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Method for DB disconnect
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStore$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStore-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataStore-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStore$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
